Move some private classes in their own files.
