+++ Src/FluentAssertions/Xml/Equivalency/XmlReaderValidator.cs
@@ -5,7 +5,7 @@ using System.Text;
 using System.Xml;
 using FluentAssertions.Execution;
 
-namespace FluentAssertions.Xml
+namespace FluentAssertions.Xml.Equivalency
 {
     internal class XmlReaderValidator
     {
@@ -213,174 +213,6 @@ namespace FluentAssertions.Xml
             }
 
             return resultBuilder.ToString();
-        }
-
-        private class ValidationResult
-        {
-            public ValidationResult(string formatString, params object[] formatParams)
-            {
-                FormatString = formatString;
-                FormatParams = formatParams;
-            }
-
-            public string FormatString { get; }
-
-            public object[] FormatParams { get; }
-        }
-
-        private class AttributeData
-        {
-            public AttributeData(string namespaceUri, string localName, string value, string prefix)
-            {
-                NamespaceUri = namespaceUri;
-                LocalName = localName;
-                Value = value;
-                Prefix = prefix;
-            }
-
-            public string NamespaceUri { get; }
-
-            public string LocalName { get; }
-
-            public string Value { get; }
-
-            private string Prefix { get; }
-
-            public string QualifiedName
-            {
-                get
-                {
-                    if (string.IsNullOrEmpty(Prefix))
-                    {
-                        return LocalName;
-                    }
-
-                    return Prefix + ":" + LocalName;
-                }
-            }
-        }
-
-        private class XmlReaderWrapper
-        {
-            private readonly XmlReader reader;
-
-            private bool skipOnce;
-
-            public XmlReaderWrapper(XmlReader reader)
-            {
-                this.reader = reader;
-
-                this.reader.MoveToContent();
-            }
-
-            public XmlNodeType NodeType => reader.NodeType;
-
-            public string LocalName => reader.LocalName;
-
-            public string NamespaceURI => reader.NamespaceURI;
-
-            public string Value => reader.Value;
-
-            public bool IsEmptyElement => reader.IsEmptyElement;
-
-            public bool EOF => reader.EOF;
-
-            public bool Read()
-            {
-                if (skipOnce)
-                {
-                    skipOnce = false;
-                    return true;
-                }
-
-                if (!reader.Read())
-                {
-                    return false;
-                }
-
-                reader.MoveToContent();
-                return true;
-            }
-
-            public void MoveToEndElement()
-            {
-                reader.Read();
-                if (reader.NodeType != XmlNodeType.EndElement)
-                {
-                    // advancing failed
-                    // skip reading on next attempt to "simulate" rewind reader
-                    skipOnce = true;
-                }
-
-                reader.MoveToContent();
-            }
-
-            public IList<AttributeData> GetAttributes()
-            {
-                var attributes = new List<AttributeData>();
-
-                if (reader.MoveToFirstAttribute())
-                {
-                    do
-                    {
-                        if (reader.NamespaceURI != "http://www.w3.org/2000/xmlns/")
-                        {
-                            attributes.Add(new AttributeData(reader.NamespaceURI, reader.LocalName, reader.Value, reader.Prefix));
-                        }
-                    }
-                    while (reader.MoveToNextAttribute());
-
-                    reader.MoveToElement();
-                }
-
-                return attributes;
-            }
-        }
-
-        private sealed class Node
-        {
-            private readonly Node parent;
-            private readonly List<Node> children = new List<Node>();
-
-            public string Name { get; }
-            public int Count { get; private set; }
-
-            private Node(Node parent, string name)
-            {
-                this.parent = parent;
-                Name = name;
-            }
-
-            public static Node CreateRoot() => new Node(null, null);
-
-            public IEnumerable<Node> GetPath()
-            {
-                Node current = this;
-                while (current.parent is object)
-                {
-                    yield return current;
-                    current = current.parent;
-                }
-            }
-
-            public Node Pop() => parent;
-
-            public Node Push(string name)
-            {
-                Node node = children.Find(e => e.Name == name)
-                    ?? AddChildNode(name);
-
-                node.Count++;
-
-                return node;
-            }
-
-            private Node AddChildNode(string name)
-            {
-                Node node = new Node(this, name);
-                children.Add(node);
-                return node;
-            }
         }
     }
 }
